requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
asyncio==4.0.0
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptg==0.5.2
cryptography==46.0.7
frozenlist==1.8.0
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
googleapis-common-protos==1.74.0
hachoir==3.3.0
httplib2==0.31.2
idna==3.11
multidict==6.7.1
oauthlib==3.3.1
propcache==0.4.1
proto-plus==1.27.2
protobuf==7.34.1
pyaes==1.6.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pyparsing==3.3.2
python-dotenv==1.2.1
requests==2.33.1
requests-oauthlib==2.0.0
rsa==4.9.1
Telethon==1.42.0
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
yarl==1.22.0
